PN-2113 varie
-modifica scenari invio b2b
-aggiunta scenario ricevente api web
-aggiunta classe per il lancio di specifici test b2b

+++ src/main/java/it/pagopa/pn/client/b2b/pa/impl/PnWebRecipientExternalClientImpl.java
@@ -56,7 +56,7 @@ public class PnWebRecipientExternalClientImpl {
     }
 
 
-    public NotificationAttachmentDownloadMetadataResponse getReceivedNotificationDocument(String iun, BigDecimal docIdx, String mandateId) throws RestClientException {
+    public NotificationAttachmentDownloadMetadataResponse getReceivedNotificationDocument(String iun, Integer docIdx, String mandateId) throws RestClientException {
         return recipientReadApi.getReceivedNotificationDocument(iun, docIdx, mandateId);
     }
